apps/server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Tour de App - backend server in Flask"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=5.0.0",
    "mysql-connector-python>=9.1.0",
    "python-dotenv>=1.0.1",
    "gunicorn>=23.0.0",
]
